+++ pw/src/main/resources/config.properties
