+++ pw/src/main/resources/database/query.sql
