+++ pw/src/main/webapp/index.jsp
@@ -1,0 +1,5 @@
+<html>
+<body>
+<h2>Hello World!</h2>
+</body>
+</html>
